+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added `mae` script
+- Added `mjustitiei` script
 
 ## [0.0.6] - 2023-03-25